Avoid calling pthread_cancel on already joined thread.
Previous implementation of _wait_for_io used pthread_join to wait for io
thread complition. In case of it taking longer than 300s we created a
_kill_thr that called pthread_cancel. However, this was called even if
pthread_join was successful which results in undefined behavior.

Instead of that the commit makes _io_thr detached and use
slurmd_cond_timedwait for 300s timeout.

Bug 16063