_handle_terminate - call step_terminate_monitor_stop() at end
The goal of step terminate monitor thread is to take action if we face
unkillable task. This thread should run for the period of time when we
attempt signaling tasks and we may potentially hang there.

Lack of the call to step_terminate_monitor_stop() results in node getting
drained even if we killed all tasks, but slurmstped continued to run for
longer time. Even if slurmstepd completes normally and there are no leftover
processes from job the node gets drained.

Bug 16063